+++ packages/tempo/src/plugins/terms/term.quarter.ts
@@ -1,4 +1,4 @@
-import { getTermRange, type Range } from './term.utils.js';
+import { defineTerm, getTermRange, type Range } from '#tempo/plugins/tempo.plugin.js';
 import { cloneify } from '#library';
 import { enums, type Tempo } from '#tempo';
 
@@ -19,17 +19,19 @@ const ranges = [
 	]
 ] as Range[][]
 
-export const key = 'qtr';
-export const scope = 'quarter';
-export const description = 'Fiscal Quarter';
+export const QuarterTerm = defineTerm({
+	key: 'qtr',
+	scope: 'quarter',
+	description: 'Fiscal Quarter',
 
-/** determine where the current Tempo instance fits within the above range */
-export function define(this: Tempo, keyOnly?: boolean) {
-	const { yy, config: { sphere } } = this;
-	const south = sphere !== COMPASS.North;										// false = North, true = South
-	const list = cloneify(ranges[+south]);										// deep clone the range
+	/** determine where the current Tempo instance fits within the above range */
+	define(this: Tempo, keyOnly?: boolean) {
+		const { yy, config: { sphere } } = this;
+		const south = sphere !== COMPASS.North;									// false = North, true = South
+		const list = cloneify(ranges[+south]);									// deep clone the range
 
-	list.forEach(itm => itm.fiscal += yy);										// calc the fiscal-year for quarter
+		list.forEach(itm => itm.fiscal += yy);									// calc the fiscal-year for quarter
 
-	return getTermRange(this, list, keyOnly);									// return the range
-}
+		return getTermRange(this, list, keyOnly);								// return the range
+	}
+});